+++ src/components/map/map-component.test.tsx
@@ -2,7 +2,7 @@ import { render, screen } from '@testing-library/react';
 import { describe, it } from 'vitest';
 import makeFakeOffer from '../../utils/make-fake-offer.ts';
 import makeFakeCityData from '../../utils/make-fake-city-data.ts';
-import Map from './map';
+import Map from './map-component.tsx';
 
 describe('Component: Map Component', () => {
   it('renders a map container correctly', () => {
