+++ src/pages/favorites/favorites-page.tsx
