+++ src/pages/login/login-page.test.tsx
@@ -1,4 +1,4 @@
-import Login from './login';
+import Login from './login-page';
 import { describe, it, vi, beforeEach } from 'vitest';
 import { render, screen, fireEvent } from '@testing-library/react';
 import { useAppDispatch, useAppSelector } from '../../hooks';
