+++ src/pages/offer/offer-page.test.tsx
@@ -1,4 +1,4 @@
-import Offer from './offer';
+import Offer from './offer-page';
 import { withStore, withHistory } from '../../utils/mocks/mock-components';
 import { screen, render } from '@testing-library/react';
 import makeFakeOffer from '../../utils/make-fake-offer';
